backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi==0.115.11
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ───────────────────────────────────────────────────────────────
sqlalchemy==2.0.39
alembic==1.15.1
asyncpg==0.30.0
geoalchemy2==0.17.1
pgvector==0.3.6

# ── Validation & settings ──────────────────────────────────────────────────
pydantic==2.12.0
pydantic-settings==2.8.1

# ── Auth ───────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── AI ─────────────────────────────────────────────────────────────────────
anthropic==0.49.0
openai==1.70.0

# ── HTTP client (webhooks, external APIs) ──────────────────────────────────
httpx==0.28.1

# ── Messaging ──────────────────────────────────────────────────────────────
resend==2.10.0

# ── PDF generation (SDG reports) ───────────────────────────────────────────
weasyprint==65.1

# ── Geospatial analytics ───────────────────────────────────────────────────
shapely==2.0.7

# ── Utilities ──────────────────────────────────────────────────────────────
python-dateutil==2.9.0
python-dotenv==1.1.0